+++ include/tmx/internal/mat/type_mat2x2.inl
@@ -33,12 +33,12 @@ namespace tmx
 
 
     template<typename T>
-    TMX_INLINE constexpr typename mat<2, 2, T>::colType& mat<2, 2, T>::operator[](size_t i) 
+    TMX_INLINE constexpr typename mat<2, 2, T>::colType& mat<2, 2, T>::operator[](int i) 
     {
         return columns[i];
     }
     template<typename T>
-    TMX_INLINE constexpr const typename mat<2, 2, T>::colType& mat<2, 2, T>::operator[](size_t i) const
+    TMX_INLINE constexpr const typename mat<2, 2, T>::colType& mat<2, 2, T>::operator[](int i) const
     {
         return columns[i];
     }
