+++ include/tmx/internal/mat/type_mat4x4.hpp
@@ -31,8 +31,8 @@ namespace tmx
         TMX_INLINE constexpr mat(const mat<4, 4, T>& m) noexcept;
 
 
-        TMX_INLINE constexpr colType& operator[](size_t i);
-        TMX_INLINE constexpr const colType& operator[](size_t i) const;
+        TMX_INLINE constexpr colType& operator[](int i);
+        TMX_INLINE constexpr const colType& operator[](int i) const;
 
 
         TMX_INLINE constexpr mat<4, 4, T>& operator=(const mat<4, 4, T>& m) noexcept = default;
